Boundary Amendments Pt2

+++ alcs-frontend/src/app/features/application/boundary-amendment/boundary-amendment.component.ts
@@ -1,10 +1,13 @@
+import { isArray } from '@angular/compiler-cli/src/ngtsc/annotations/common';
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
+import moment from 'moment/moment';
 import { Subject, takeUntil } from 'rxjs';
 import { ApplicationBoundaryAmendmentDto } from '../../../services/application/application-boundary-amendments/application-boundary-amendment.dto';
 import { ApplicationBoundaryAmendmentService } from '../../../services/application/application-boundary-amendments/application-boundary-amendment.service';
 import { ApplicationDetailService } from '../../../services/application/application-detail.service';
 import { ApplicationDto } from '../../../services/application/application.dto';
+import { ConfirmationDialogService } from '../../../shared/confirmation-dialog/confirmation-dialog.service';
 import { EditBoundaryAmendmentDialogComponent } from './edit-boundary-amendment-dialog/edit-boundary-amendment-dialog.component';
 
 @Component({
@@ -17,10 +20,30 @@ export class BoundaryAmendmentComponent implements OnInit, OnDestroy {
   application?: ApplicationDto;
   amendments: ApplicationBoundaryAmendmentDto[] = [];
   private fileNumber: string = '';
+  years: { label: string; value: string }[] = [];
+  periods: { label: string; value: string }[] = [
+    {
+      label: '1',
+      value: '1',
+    },
+    {
+      label: '2',
+      value: '2',
+    },
+    {
+      label: '3',
+      value: '3',
+    },
+    {
+      label: '4',
+      value: '4',
+    },
+  ];
 
   constructor(
     private boundaryAmendmentService: ApplicationBoundaryAmendmentService,
     private applicationDetailService: ApplicationDetailService,
+    private confirmationDialogService: ConfirmationDialogService,
     private dialog: MatDialog,
   ) {}
 
@@ -32,6 +55,14 @@ export class BoundaryAmendmentComponent implements OnInit, OnDestroy {
         this.loadAmendments();
       }
     });
+
+    const currentYear = moment().year();
+    for (let i = currentYear; i >= 1974; i--) {
+      this.years.push({
+        label: i.toString(10),
+        value: i.toString(10),
+      });
+    }
   }
 
   async loadAmendments() {
@@ -77,7 +108,33 @@ export class BoundaryAmendmentComponent implements OnInit, OnDestroy {
   }
 
   onDeleteAmendment(uuid: string) {
-    this.boundaryAmendmentService.delete(uuid);
-    this.loadAmendments();
+    this.confirmationDialogService
+      .openDialog({
+        body: 'Are you sure you want to delete the selected Amendment?',
+      })
+      .subscribe(async (didConfirm) => {
+        if (didConfirm) {
+          await this.boundaryAmendmentService.delete(uuid);
+          await this.loadAmendments();
+        }
+      });
+  }
+
+  async onSaveYear(uuid: string, $event: string | string[] | null) {
+    if ($event && !Array.isArray($event)) {
+      await this.boundaryAmendmentService.update(uuid, {
+        year: parseInt($event),
+      });
+      await this.loadAmendments();
+    }
+  }
+
+  async onSavePeriod(uuid: string, $event: string | string[] | null) {
+    if ($event && !Array.isArray($event)) {
+      await this.boundaryAmendmentService.update(uuid, {
+        period: parseInt($event),
+      });
+      await this.loadAmendments();
+    }
   }
 }
